refactor(tools): extract card-metadata-tools module (#66)
Architectural improvements from code review:
- Extract 5 read-only metadata ops to card-metadata-tools.ts
- Move getCardSize from move-tools (semantic correction)
- Move getCardCustomFields/Types/History/Outcomes from crud-tools

LOC improvements:
- card-crud-tools.ts: 385 → 266 (6% over target vs 54% over)
- card-move-tools.ts: 124 → 82 (write-only module now)
- NEW card-metadata-tools.ts: 175 (read-only queries)

Also adds 4 tests for bulk operation edge cases:
- Partial success handling for bulk delete
- All failures handling for bulk delete
- Partial success handling for bulk update
- All failures handling for bulk update
