+++ src/server/tools/card-tools/index.ts
@@ -4,6 +4,7 @@ import { BusinessMapClientFactory } from '@client/client-factory.js';
 import { BaseToolHandler } from '../base-tool.js';
 import { registerCardCrudTools } from './card-crud-tools.js';
 import { registerCardMoveTools } from './card-move-tools.js';
+import { registerCardMetadataTools } from './card-metadata-tools.js';
 import { registerCardCommentTools } from './card-comment-tools.js';
 import { registerCardSubtaskTools } from './card-subtask-tools.js';
 import { registerCardRelationshipTools } from './card-relationship-tools.js';
@@ -12,8 +13,9 @@ import { registerCardBulkTools } from './card-bulk-tools.js';
 /**
  * CardToolHandler orchestrates registration of all card-related MCP tools.
  * Tools are organized into focused modules by domain:
- * - CRUD: list, get, create, update, delete cards + metadata
- * - Move: move card, get/set card size
+ * - CRUD: list, get, create, update, delete cards
+ * - Move: move card, set card size (write operations)
+ * - Metadata: card size, custom fields, types, history, outcomes (read queries)
  * - Comments: comment CRUD operations
  * - Subtasks: subtask operations
  * - Relationships: parents, children, linked cards
@@ -28,6 +30,7 @@ export class CardToolHandler implements BaseToolHandler {
   ): void {
     registerCardCrudTools(server, clientOrFactory, readOnlyMode, enabledTools);
     registerCardMoveTools(server, clientOrFactory, readOnlyMode, enabledTools);
+    registerCardMetadataTools(server, clientOrFactory, enabledTools);
     registerCardCommentTools(server, clientOrFactory, readOnlyMode, enabledTools);
     registerCardSubtaskTools(server, clientOrFactory, readOnlyMode, enabledTools);
     registerCardRelationshipTools(server, clientOrFactory, readOnlyMode, enabledTools);
